Print input and output file sizes (#914)

+++ src/archive/tar.rs
@@ -25,7 +25,7 @@ use crate::{
 
 /// Unpacks the archive given by `archive` into the folder given by `into`.
 /// Assumes that output_folder is empty
-pub fn unpack_archive(reader: impl Read, output_folder: &Path) -> crate::Result<usize> {
+pub fn unpack_archive(reader: impl Read, output_folder: &Path) -> crate::Result<u64> {
     let mut archive = tar::Archive::new(reader);
 
     let mut files_unpacked = 0;
